쿼리 테스트 완료

+++ src/main/java/study/thboard2/domain/vo/BoardVo.java
@@ -5,7 +5,7 @@ import lombok.Data;
 @Data
 public class BoardVo {
 
-    private Long boardNo;                       //게시글 순번
+    private Integer boardNo;                       //게시글 순번
     private String title;                       //제목
     private String writer;                      //작성자
     private String content;                     //내용
